seperated logic for /paper/count endpoint
