+++ src/app/api/papers/count/route.ts
@@ -1,21 +1,13 @@
 import { NextResponse } from "next/server";
-import { connectToDatabase } from "@/lib/database/mongoose";
-import CourseCount from "@/db/course";
+import { getCourseCounts } from "@/lib/services/paper";
 
 export const dynamic = "force-dynamic";
 
 export async function GET(req: Request) {
   try {
-    await connectToDatabase();
+    const courseCount = await getCourseCounts();
 
-    const count = await CourseCount.find().lean();
-
-    const formatted = count.map((item) => ({
-      name: item.name,
-      count: item.count,
-    }));
-
-    return NextResponse.json(formatted, { status: 200 });
+    return NextResponse.json(courseCount, { status: 200 });
   } catch (error) {
     return NextResponse.json(
       { message: "Failed to fetch course counts", error },
