+++ src/lib/services/paper.ts
@@ -3,6 +3,7 @@ import { type IPaper } from "@/interface";
 import { escapeRegExp } from "@/lib/utils/regex";
 import { extractUniqueValues } from "@/lib/utils/paper-aggregation";
 import { connectToDatabase } from "../database/mongoose";
+import CourseCount from "@/db/course";
 
 export async function getPapersBySubject(subject: string) {
     if (!subject){
@@ -33,4 +34,17 @@ export async function getPaperById(id: string) {
     }
 
     return paper;
+}
+
+export async function getCourseCounts(){
+    await connectToDatabase();
+
+    const count = await CourseCount.find().lean();
+
+    const formatted = count.map((item) => ({
+        name: item.name,
+        count: item.count,
+    }));
+
+    return formatted;
 }
